+++ NEWS.md
@@ -10,6 +10,8 @@
 - Drop missing values and reduce size of `gss` dataset used in examples (#282)
 - Fix formatting of lifecycle badges (#283)
 - Add `stat = "ratio of props"` and `stat = "odds ratio"` to `calculate` (#285)
+- Add `prop_test()`, a tidy interface to `prop.test()` (#284, #287)
+- Updates to `visualize()` for compatibility with `ggplot2` v3.3.0 (#289)
 
 # infer 0.5.1